Comitting my current state of converting single.txt over to both a trie and an array.
Have not tried loading these in browsers yet. Will build some test pages and see how
they work in a future check-in. To properly load a dictionary file is likely going to
require a lot more compression.

One optimization I made was to save off 3 characters as the initial discriminator. So
basically a prefix discriminator. This could likely be encoded as a single, unrecognizable
character instead.

0-25 = 5 bits, 15 bits < short. We could use fromCharCode http://es5.github.com/#x15.5.3.2
to make that a single character. Depending on the algorithm processing the dictionary computing
the triplets as the first order of business might be a minimal tax to pay. Note it requires
saving two character strings as a "{char1}{char2}{null}" and one character strings as "{char1}{null}{null}"
which isn't a terrible issue since the number of 1 and 2 letter words is by combinators small (26 + 26^2 or
702 entries maximum)